moved contents of assessment package to exam package
