+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/Assessment.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment;
+package org.opentestsystem.rdw.reporting.exam;
 
 import com.google.common.collect.ImmutableList;
