+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/DefaultExamItemService.java
@@ -3,7 +3,7 @@ package org.opentestsystem.rdw.reporting.exam;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.io.ByteStreams;
-import org.opentestsystem.rdw.reporting.assessment.artifact.ArtifactRepository;
+import org.opentestsystem.rdw.reporting.exam.artifact.ArtifactRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
