+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/DefaultGroupExamService.java
@@ -1,8 +1,7 @@
-package org.opentestsystem.rdw.reporting.assessment;
+package org.opentestsystem.rdw.reporting.exam;
 
 import com.google.common.collect.ImmutableList;
-import org.opentestsystem.rdw.reporting.exam.Exam;
-import org.opentestsystem.rdw.reporting.exam.GroupExam;
+import org.opentestsystem.rdw.reporting.exam.repository.AssessmentRepository;
 import org.opentestsystem.rdw.reporting.exam.repository.GroupExamRepository;
 import org.opentestsystem.rdw.reporting.exam.repository.GroupIabExamRepository;
 import org.opentestsystem.rdw.reporting.exam.repository.GroupIcaSummativeExamRepository;
@@ -20,7 +19,7 @@ import static com.google.common.collect.Lists.newArrayList;
 import static java.util.Comparator.comparing;
 
 @Service
-public class DefaultGroupAssessmentService implements GroupAssessmentService {
+public class DefaultGroupExamService implements GroupExamService {
 
     private static final long IAB_TYPE_ID = 2L;
     private static final Comparator<GroupExam> groupExamCompletedDateTimeComparator = comparing(GroupExam::getDateTime);
@@ -30,7 +29,7 @@ public class DefaultGroupAssessmentService implements GroupAssessmentService {
 
 
     @Autowired
-    public DefaultGroupAssessmentService(
+    public DefaultGroupExamService(
             @NotNull final GroupIabExamRepository groupInterimAssessmentRepository,
             @NotNull final GroupIcaSummativeExamRepository groupIcaSummativeExamRepository,
             @NotNull final AssessmentRepository assessmentRepository) {
