+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/GroupExam.java
@@ -1,7 +1,6 @@
 package org.opentestsystem.rdw.reporting.exam;
 
 import com.google.common.collect.ImmutableList;
-import org.opentestsystem.rdw.reporting.assessment.Assessment;
 
 import java.time.Instant;
 import java.util.List;
