+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/GroupExamService.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment;
+package org.opentestsystem.rdw.reporting.exam;
 
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.GroupExam;
@@ -12,7 +12,7 @@ import java.util.Optional;
 /**
  * Service responsible for managing assessments taken by a group
  */
-public interface GroupAssessmentService {
+public interface GroupExamService {
 
     /**
      * Gets the most recent assessment taken by a group for a given school year.
