+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/Subjects.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment;
+package org.opentestsystem.rdw.reporting.exam;
 
 /**
  * Utility class intended for support in handling assessment subjects
