+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/artifact/AmazonS3ArtifactRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment.artifact;
+package org.opentestsystem.rdw.reporting.exam.artifact;
 
 import com.amazonaws.services.s3.model.AmazonS3Exception;
 import org.springframework.beans.factory.annotation.Autowired;
