+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/artifact/ArtifactRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment.artifact;
+package org.opentestsystem.rdw.reporting.exam.artifact;
 
 import java.io.IOException;
 import java.io.InputStream;
