+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/artifact/ArtifactRepositorySettings.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment.artifact;
+package org.opentestsystem.rdw.reporting.exam.artifact;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
