+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/repository/AbstractJdbcGroupExamRepository.java
@@ -2,13 +2,12 @@ package org.opentestsystem.rdw.reporting.exam.repository;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import org.opentestsystem.rdw.reporting.assessment.Assessment;
+import org.opentestsystem.rdw.reporting.exam.Assessment;
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.GroupExam;
 import org.opentestsystem.rdw.reporting.exam.Student;
 import org.opentestsystem.rdw.reporting.exam.StudentContext;
 import org.opentestsystem.rdw.reporting.security.GroupSecurityContext;
-import org.springframework.dao.DataAccessException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
