+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/repository/AssessmentRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment;
+package org.opentestsystem.rdw.reporting.exam.repository;
 
 public interface AssessmentRepository {
