+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/repository/GroupExamRepository.java
@@ -1,13 +1,12 @@
 package org.opentestsystem.rdw.reporting.exam.repository;
 
-import org.opentestsystem.rdw.reporting.assessment.Assessment;
+import org.opentestsystem.rdw.reporting.exam.Assessment;
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.GroupExam;
 import org.opentestsystem.rdw.reporting.security.GroupSecurityContext;
 
 import javax.validation.constraints.NotNull;
 import java.util.List;
-import java.util.Optional;
 
 /**
  * Interface for accessing group exams
