+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/repository/JdbcAssessmentRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.assessment;
+package org.opentestsystem.rdw.reporting.exam.repository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
