+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/repository/JdbcGroupIabExamRepository.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.reporting.exam.repository;
 
-import org.opentestsystem.rdw.reporting.assessment.Assessment;
+import org.opentestsystem.rdw.reporting.exam.Assessment;
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.ScaleScore;
 import org.springframework.beans.factory.annotation.Value;
