+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/repository/JdbcGroupIcaSummativeExamRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.reporting.exam.repository;
 
 import com.google.common.collect.ImmutableList;
-import org.opentestsystem.rdw.reporting.assessment.Assessment;
+import org.opentestsystem.rdw.reporting.exam.Assessment;
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.ScaleScore;
 import org.springframework.beans.factory.annotation.Value;
