+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/group/JdbcGroupRepository.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.reporting.group;
 
-import org.opentestsystem.rdw.reporting.assessment.Subjects;
+import org.opentestsystem.rdw.reporting.exam.Subjects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
