+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/web/GroupAssessmentController.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.reporting.web;
 
-import org.opentestsystem.rdw.reporting.assessment.Assessment;
-import org.opentestsystem.rdw.reporting.assessment.GroupAssessmentService;
+import org.opentestsystem.rdw.reporting.exam.Assessment;
+import org.opentestsystem.rdw.reporting.exam.GroupExamService;
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.GroupExam;
 import org.opentestsystem.rdw.reporting.security.User;
@@ -19,10 +19,10 @@ import java.util.NoSuchElementException;
 @RestController
 public class GroupAssessmentController {
 
-    private final GroupAssessmentService service;
+    private final GroupExamService service;
 
     @Autowired
-    public GroupAssessmentController(@NotNull final GroupAssessmentService service) {
+    public GroupAssessmentController(@NotNull final GroupExamService service) {
         this.service = service;
     }
