+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/exam/DefaultExamItemServiceTest.java
@@ -4,9 +4,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
-import org.opentestsystem.rdw.reporting.exam.DefaultExamItemService;
-import org.opentestsystem.rdw.reporting.exam.ScoringCriterion;
-import org.opentestsystem.rdw.reporting.assessment.artifact.ArtifactRepository;
+import org.opentestsystem.rdw.reporting.exam.artifact.ArtifactRepository;
 
 import javax.imageio.ImageIO;
 import java.awt.Graphics;
