+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/exam/artifact/ArtifactRepositoryIT.java
@@ -1,11 +1,8 @@
-package org.opentestsystem.rdw.reporting.assessment.artifact;
+package org.opentestsystem.rdw.reporting.exam.artifact;
 
 import com.amazonaws.util.IOUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.opentestsystem.rdw.reporting.assessment.artifact.AmazonS3ArtifactRepository;
-import org.opentestsystem.rdw.reporting.assessment.artifact.ArtifactRepository;
-import org.opentestsystem.rdw.reporting.assessment.artifact.ArtifactRepositorySettings;
 import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
